+++ aeron-client/src/main/c/aeron_client_conductor.c
@@ -370,7 +370,7 @@ void aeron_client_conductor_on_driver_response(int32_t type_id, uint8_t *buffer,
     return;
 
 malformed_command:
-    AERON_CLIENT_FORMAT_BUFFER(error_message, "command=%x too short: length=%" PRIu32, type_id, (uint32_t)length);
+    AERON_CLIENT_FORMAT_BUFFER(error_message, "command=%x too short: length=%" PRIu64, type_id, (uint64_t)length);
     conductor->error_handler(conductor->error_handler_clientd, AERON_ERROR_CODE_MALFORMED_COMMAND, error_message);
 }
